+++ Wox.Infrastructure/Wox.Infrastructure.csproj
@@ -52,6 +52,7 @@
     <Compile Include="CommonStorage.cs" />
     <Compile Include="FuzzyMatcher.cs" />
     <Compile Include="GloablHotkey.cs" />
+    <Compile Include="GlobalHotkey.cs" />
     <Compile Include="HotkeyModel.cs" />
     <Compile Include="IniParser.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
